Update to v2.5.2: * More fixes for issue #78: StepAddInc (DeltaR) switched from signed 16 to signed 32 bit value in FIFO data structure. * Also updated release notes and EBB docs to reflect this change.

+++ EBB_firmware/app.X/source/ebb.h
@@ -80,7 +80,7 @@ typedef struct
 {
     CommandType     Command;
     INT32           StepAdd[NUMBER_OF_STEPPERS];
-    INT16           StepAddInc[NUMBER_OF_STEPPERS];
+    INT32           StepAddInc[NUMBER_OF_STEPPERS];
     UINT32          StepsCounter[NUMBER_OF_STEPPERS];
     UINT8           DirBits;
     UINT32          DelayCounter;   // NOT Milliseconds! In 25KHz units
